% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-operations.R
\name{emat-operations}
\alias{emat-operations}
\alias{emat_compare}
\alias{emat_complement}
\alias{emat_sort}
\alias{order_rows}
\alias{emat_saturated_model}
\alias{emat_loop_model0}
\alias{emat_random_tree0}
\alias{emat_loop_model}
\alias{emat_random_tree}
\alias{emat_line_model}
\alias{emat_star_model}
\alias{emat_grid_model}
\alias{emat_rectangular_grid}
\alias{emat_random_model}
\alias{emat_cq_random_model}
\title{Edge matrix operations}
\usage{
emat_compare(emat1, emat2)

emat_complement(emat1, emat2)

emat_sort(emat1)

order_rows(emat)

emat_saturated_model(index)

emat_loop_model0(index)

emat_random_tree0(index)

emat_loop_model(index, prob = 0)

emat_random_tree(index, prob = 0)

emat_line_model(index)

emat_star_model(index)

emat_grid_model(nr.nc)

emat_rectangular_grid(dim)

emat_random_model(index, prob = 0.1)

emat_cq_random_model(index, prob = 0.1)
}
\arguments{
\item{emat, emat1, emat2}{Edge matrix (a 2 x p matrix)}

\item{index}{A vector of integers}

\item{prob}{Probability of any edge being present.}

\item{nr.nc}{Vector with number of rows and columns.}

\item{dim}{A vector with dimensions}
}
\description{
Edge matrix operations needed for ips algorithms
}
\details{
An emat with p edges is represented by a 2 x p matrix.
}
\note{
These functions may well be removed from the package in future relases
}
\examples{
emat1 <- emat_saturated_model(3:4)
emat2 <- emat_saturated_model(1:4)
emat_complement(emat1, emat2)
emat3 <- emat_saturated_model(2:4)
emat_compare(emat1, emat3)
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
