% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ggm.R
\name{fit_ggm}
\alias{fit_ggm}
\title{Fit Gaussian graphical models}
\usage{
fit_ggm(
  S,
  edges = NULL,
  nobs,
  K = NULL,
  maxit = 10000L,
  eps = 1e-06,
  convcrit = 1,
  aux = list(),
  method = "covips",
  print = 0
)
}
\arguments{
\item{S}{Sample covariance matrix.}

\item{edges}{Generators of model; a list of integer vectors or a 2
x p matrix of integers.}

\item{nobs}{Number of observations}

\item{K}{Initial value of concentration matrix.}

\item{maxit}{Maximum number of iterations.}

\item{eps}{Convergence criterion.}

\item{convcrit}{Convergence criterions. See section `details`.}

\item{aux}{A list of form name=value.}

\item{method}{Either `"ips"` or `"fips"`.}

\item{print}{Should output from fitting be printed?}
}
\description{
Fit Gaussian graphical models using various algorithms.
}
\details{
Convergence criterion:

* 1: max absolute difference between S and Sigmahat on edges.

* 2: difference in log likelihood divided by number of parameters in the
model (number of edges + number of nodes) between successive
iterations.

R-based / c++-based in combination with con / cov.
}
\examples{
options("digits"=3)
data(math)

S <- cov(math)
nobs <- nrow(math)
gl <- list(1:3, 3:5)
em <- matrix(c(1,2, 2,3, 1,3, 3,4, 3,5, 4,5), nrow=2)

EPS = 1e-2

cg = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="cov")
ci = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="con")
rg = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="cov", aux=list(engine="R"))
ri = fit_ggm(S, gl, nobs=nobs, eps=EPS, method="con", aux=list(engine="R"))

K <- solve(S)
(ci$K - K)  \%>\% abs \%>\% max
(cg$K - K)  \%>\% abs \%>\% max
(ri$K - K)  \%>\% abs \%>\% max
(rg$K - K)  \%>\% abs \%>\% max

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
