% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-grips.R
\name{utes}
\alias{utes}
\alias{does_fit}
\title{Utes}
\usage{
does_fit(Sigma, S, emat, eps = 1e-04)
}
\arguments{
\item{Sigma, S}{Symmetrix positive definite matrices}

\item{emat}{Edge matrix (2 x p)}

\item{eps}{Threshold}
}
\description{
Utes
}
