% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_grips.R
\name{parmEq}
\alias{parmEq}
\alias{utilities}
\alias{as_emat2cq}
\alias{as_emat_complement}
\alias{as_emat2amat}
\alias{as_emat2elist}
\alias{as_elist2emat}
\alias{as_glist2emat}
\alias{as_glist2cq}
\alias{as_glist2graph}
\alias{as_glist2igraph}
\alias{as_emat2graph}
\alias{as_emat2igraph}
\alias{as_amat2emat}
\alias{as_emat2glist}
\alias{as_glist2out_edges}
\alias{as_K2amat}
\alias{as_K2graph}
\alias{as_sparse}
\alias{logLik.gips_fit_class}
\alias{AIC.gips_fit_class}
\alias{BIC.gips_fit_class}
\alias{sigma.gips_fit_class}
\alias{concentration}
\alias{concentration.gips_fit_class}
\alias{print.gips_fit_class}
\alias{summary.gips_fit_class}
\alias{glance.gips_fit_class}
\title{Parameter equality}
\usage{
parmEq(p2a, p2b, eps = 1e-04)

as_emat2cq(emat, nvar = NULL)

as_emat_complement(emat, nvar)

as_emat2amat(emat, d)

as_emat2elist(emat)

as_elist2emat(elist)

as_glist2emat(glist)

as_glist2cq(glist)

as_glist2graph(glist, d)

as_glist2igraph(glist, d)

as_emat2graph(emat, d)

as_emat2igraph(emat, d)

as_amat2emat(amat, eps = 1e-04)

as_emat2glist(emat)

as_glist2out_edges(glist)

as_K2amat(K, eps = 1e-04)

as_K2graph(K)

as_sparse(K)

\method{logLik}{gips_fit_class}(object, ...)

\method{AIC}{gips_fit_class}(object, ..., k = 2)

\method{BIC}{gips_fit_class}(object, ...)

\method{sigma}{gips_fit_class}(object, ...)

concentration(object, ...)

\method{concentration}{gips_fit_class}(object, ...)

\method{print}{gips_fit_class}(x, ...)

\method{summary}{gips_fit_class}(object, ...)

glance.gips_fit_class(x, ...)
}
\arguments{
\item{p2a, p2b}{Parameter lists.}

\item{eps}{A small number.}

\item{emat}{Edge matrix (2 x p)}

\item{nvar}{Number of variables}

\item{d}{Number of columns in output.

glist <- list(c(1,2,3),c(2,3,4),c(5,6))
em <- as_glist2emat(glist)
am <- as_emat2amat(em, d=6)
ig <- as_emat2igraph(em)
el <- as_emat2elist(em)
igraph::max_cliques(ig)
as_emat2cq(em, 6)
as_emat_complement(em, 6)}

\item{elist}{Edge list (list of pairs)}

\item{glist}{Generator list}

\item{amat}{Adjacency matrix}

\item{K}{Concentration matrix}

\item{object}{Model object.}

\item{...}{Additional arguments; currently not used.}

\item{k}{Penalty parameter for calculating AIC; only k=2 gives genuine AIC.}

\item{x}{An object (to print)}
}
\description{
Utilities
}
